[quant][refactor tests] Move test_quantized_functional.py into test_quantize_eager_ptq.py Summary:
Test Plan:
python test/test_quantization.py TestQuantizedFunctionalOps

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
