+++ test/test_quantization.py
@@ -16,8 +16,6 @@ from quantization.test_quantized_op import TestPadding  # noqa: F401
 from quantization.test_quantized_op import TestQuantizedEmbeddingOps  # noqa: F401
 from quantization.test_quantized_op import TestDynamicQuantizedRNNOp  # noqa: F401
 
-# Quantized Functional
-from quantization.test_quantized_functional import TestQuantizedFunctional  # noqa: F401
 
 # Quantized Module
 from quantization.test_quantized_module import TestStaticQuantizedModule  # noqa: F401
@@ -45,6 +43,7 @@ from quantization.test_quantize_eager_ptq import TestEagerModeActivationOps  # n
 from quantization.test_quantize_eager_ptq import TestFunctionalModule  # noqa: F401
 from quantization.test_quantize_eager_ptq import TestModelNumerics  # noqa: F401
 from quantization.test_quantize_eager_ptq import TestQuantizeONNXExport  # noqa: F401
+from quantization.test_quantize_eager_ptq import TestQuantizedFunctionalOps  # noqa: F401
 
 # Eager mode quantization aware training
 from quantization.test_quantize_eager_qat import TestQuantizationAwareTraining  # noqa: F401
